Add new ListCallback class to avoid code duplication + refactor classes
