+++ app/src/main/java/com/github/brugapp/brug/ui/ChatMenuActivity.kt
@@ -16,10 +16,13 @@ import com.github.brugapp.brug.ui.components.BottomNavBar
 import com.github.brugapp.brug.ui.components.CustomTopBar
 import com.github.brugapp.brug.view_model.ChatListAdapter
 import com.github.brugapp.brug.view_model.ChatMenuViewModel
+import com.github.brugapp.brug.view_model.ListCallback
 import com.google.android.material.snackbar.Snackbar
 
 private const val DUMMY_TEXT: String = "Actual behavior coming soon…"
 private const val SEARCH_HINT: String = "Search for a conversation…"
+private const val CHECK_TEXT: String = "The conversation has been marked as resolved."
+
 
 class ChatMenuActivity : AppCompatActivity() {
 
@@ -55,11 +58,20 @@ class ChatMenuActivity : AppCompatActivity() {
 
         listView.layoutManager = LinearLayoutManager(this)
 
+        val dragPair = Pair(
+            0, ItemTouchHelper.LEFT.or(ItemTouchHelper.RIGHT)
+        )
+
         val swipePair = Pair(
             ContextCompat.getDrawable(this, R.drawable.ic_baseline_check_circle_outline_24)!!,
             ContextCompat.getColor(this, R.color.chat_list_resolve_BG))
 
-        val listCallback = model.ChatListCallback(swipePair, swipePair, listViewAdapter)
+        val listAdapterPair = Pair(
+            model.getChatList(),
+            listViewAdapter
+        )
+
+        val listCallback = ListCallback(CHECK_TEXT, dragPair, swipePair, listAdapterPair)
         ItemTouchHelper(listCallback).attachToRecyclerView(listView)
         listView.adapter = listViewAdapter
